Create the functions to convert from any category
Initially, it was supposed to make files with translation for each
category, but in this case it would be simple to 'copy' the code from
file to file. I managed to implement a function that is suitable for any
category, if there are text files for it with data for translation.
Therefore, the file itself with the implementation of functions's
renamed from 'timeunit.c' to 'units.c'